The Article page is divided into components: ArticleImage and ArticleInfo
