+++ src/components/ads/ArticleInfo/ArticleInfo.module.css
@@ -1,0 +1,164 @@
+.articleBlock {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: column;
+    flex-direction: column;
+}
+
+.articleTitle {
+    margin-bottom: 10px;
+    font-size: 32px;
+    line-height: 46px;
+    font-weight: 500;
+    color: #000000;
+}
+
+.articleInfo {
+    margin-bottom: 34px;
+}
+
+.articleDate,
+.articleCity {
+    font-size: 16px;
+    line-height: 21px;
+    color: #5f5f5f;
+    margin-bottom: 4px;
+}
+
+.articleLink {
+    font-size: 16px;
+    line-height: 21px;
+    color: #009ee4;
+}
+
+.articlePrice {
+    font-size: 28px;
+    line-height: 39px;
+    font-weight: 700;
+    margin-bottom: 20px;
+}
+
+.btnBlock {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -ms-flex-wrap: wrap;
+    flex-wrap: wrap;
+    flex-direction: row;
+    gap: 10px;
+}
+
+.articleAuthor {
+    margin-top: 24px;
+    margin-bottom: 20px;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: horizontal;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: row;
+    flex-direction: row;
+    -webkit-box-align: start;
+    -ms-flex-align: start;
+    align-items: flex-start;
+}
+
+.authorImg {
+    width: 40px;
+    height: 40px;
+    background-color: #f0f0f0;
+    border-radius: 50%;
+}
+
+.authorImg img {
+    width: 100%;
+    height: auto;
+    display: block;
+    -o-object-fit: cover;
+    object-fit: cover;
+}
+
+.authorCont {
+    margin-left: 12px;
+}
+
+.authorName {
+    font-size: 20px;
+    line-height: 26px;
+    font-weight: 600;
+    color: #009ee4;
+}
+
+.authorAbout {
+    font-size: 16px;
+    line-height: 32px;
+    color: #5f5f5f;
+}
+.span {
+    display: block;
+    font-size: 14px;
+    font-weight: 400;
+}
+
+@media screen and (max-width: 966px) {
+    .articleBtn {
+        width: 225px;
+        margin-right: 0;
+    }
+    .btnRemove {
+        width: 225px;
+    }
+}
+
+@media screen and (max-width: 768px) {
+    .articleTitle {
+        margin-bottom: 10px;
+        font-size: 18px;
+        line-height: 1;
+    }
+    .articleInfo {
+        margin-bottom: 20px;
+    }
+    .articleDate,
+    .articleCity {
+        font-size: 14px;
+        line-height: 17px;
+        color: #5f5f5f;
+        margin-bottom: 4px;
+    }
+    .articleLink {
+        font-size: 14px;
+        line-height: 19px;
+        color: #009ee4;
+    }
+    .articlePrice {
+        font-size: 18px;
+        line-height: 25px;
+        font-weight: 700;
+        margin-bottom: 20px;
+    }
+    .articleBtn {
+        width: 100%;
+        height: 57px;
+        font-size: 14px;
+        line-height: 20px;
+        color: #ffffff;
+    }
+
+    .articleAuthor {
+        margin-top: 30px;
+        margin-bottom: 30px;
+    }
+    .authorName {
+        font-size: 18px;
+        line-height: 23px;
+        font-weight: 600;
+    }
+    .authorAbout {
+        font-size: 14px;
+        line-height: 28px;
+    }
+}
