+++ src/pages/Article/Article.module.css
@@ -34,10 +34,6 @@
 .mainTitle {
     margin-bottom: 32px;
     padding: 0 5px;
-}
-
-.mainTitle,
-.articleTitle {
     font-size: 32px;
     line-height: 46px;
     font-weight: 500;
@@ -74,17 +70,6 @@
     display: flex;
 }
 
-.menu__btn-serch {
-    width: 241px;
-    height: 50px;
-    background-color: #009ee4;
-    border: 1px solid #009ee4;
-    color: #ffffff;
-    border-radius: 6px;
-    font-size: 16px;
-    line-height: 24px;
-}
-
 .articContent {
     display: -webkit-box;
     display: -ms-flexbox;
@@ -99,207 +84,13 @@
     max-width: 480px;
 }
 
-.articleFillImg {
-    width: 100%;
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-orient: vertical;
-    -webkit-box-direction: normal;
-    -ms-flex-direction: column;
-    flex-direction: column;
-    -webkit-box-align: center;
-    -ms-flex-align: center;
-    align-items: center;
-}
-
-.articleImg {
-    width: 480px;
-    height: 480px;
-    background-color: #f0f0f0;
-    margin: 0 5px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.articleImg img {
-    width: 100%;
-    height: 100%;
-    display: block;
-    -o-object-fit: cover;
-    object-fit: cover;
-    cursor: pointer;
-}
-
-.articleImg img:hover {
-    border: 2px solid #009ee4;
-}
-
-.imgBarMob {
-    display: none;
-}
-
-.imgBar {
-    margin-top: 30px;
-    width: 490px;
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-orient: horizontal;
-    -webkit-box-direction: normal;
-    -ms-flex-direction: row;
-    flex-direction: row;
-    -webkit-box-align: center;
-    -ms-flex-align: center;
-    align-items: center;
-    -webkit-box-pack: left;
-    -ms-flex-pack: left;
-    justify-content: left;
-    overflow: hidden;
-    margin-left: -5px;
-}
-
-.imgBarDiv {
-    width: 88px;
-    min-width: 88px;
-    height: 88px;
-    background-color: #f0f0f0;
-    border: 2px solid #f0f0f0;
-    margin: 0 5px;
-    transition: all 0.3s;
-}
-.imgBarDiv img {
-    width: 100%;
-    height: 100%;
-    display: block;
-    -o-object-fit: cover;
-    object-fit: cover;
-    cursor: pointer;
-}
-.imgBarDiv:hover {
-    border: 2px solid #009ee4;
-}
-
 .articleRight {
     max-width: 621px;
-}
-
-.articleBlock {
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-orient: vertical;
-    -webkit-box-direction: normal;
-    -ms-flex-direction: column;
-    flex-direction: column;
-}
-
-.articleTitle {
-    margin-bottom: 10px;
-}
-
-.articleInfo {
-    margin-bottom: 34px;
-}
-
-.articleDate,
-.articleCity {
-    font-size: 16px;
-    line-height: 21px;
-    color: #5f5f5f;
-    margin-bottom: 4px;
-}
-
-.articleLink {
-    font-size: 16px;
-    line-height: 21px;
-    color: #009ee4;
-}
-
-.articlePrice {
-    font-size: 28px;
-    line-height: 39px;
-    font-weight: 700;
-    margin-bottom: 20px;
-}
-
-.btnBlock {
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -ms-flex-wrap: wrap;
-    flex-wrap: wrap;
-    flex-direction: row;
-    gap: 10px;
-}
-
-.articleAuthor {
-    margin-top: 24px;
-    margin-bottom: 20px;
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-orient: horizontal;
-    -webkit-box-direction: normal;
-    -ms-flex-direction: row;
-    flex-direction: row;
-    -webkit-box-align: start;
-    -ms-flex-align: start;
-    align-items: flex-start;
-}
-
-.authorImg {
-    width: 40px;
-    height: 40px;
-    background-color: #f0f0f0;
-    border-radius: 50%;
-}
-
-.authorImg img {
-    width: 100%;
-    height: auto;
-    display: block;
-    -o-object-fit: cover;
-    object-fit: cover;
-}
-
-.authorCont {
-    margin-left: 12px;
-}
-
-.authorName {
-    font-size: 20px;
-    line-height: 26px;
-    font-weight: 600;
-    color: #009ee4;
-}
-
-.authorAbout {
-    font-size: 16px;
-    line-height: 32px;
-    color: #5f5f5f;
-}
-.noPhoto {
-    color: #5f5f5f;
-    font-size: 22px;
-}
-.span {
-    display: block;
-    font-size: 14px;
-    font-weight: 400;
 }
 
 @media screen and (max-width: 966px) {
     .articleLeft {
         margin-right: 20px;
-    }
-    .articleBtn {
-        width: 225px;
-        margin-right: 0;
-    }
-    .btnRemove {
-        width: 225px;
     }
     .articContent {
         -webkit-box-pack: center;
@@ -324,12 +115,10 @@
     .mainTitle {
         margin-bottom: 14px;
         padding: 0;
-    }
-    .mainTitle,
-    .articleTitle {
         font-size: 18px;
         line-height: 1;
     }
+
     .mainContent {
         width: 100%;
         margin: 0 auto;
@@ -367,138 +156,10 @@
         margin-right: 0;
         margin-bottom: 20px;
     }
-    .articleFillImg {
-        position: relative;
-        min-width: 320px;
-        min-height: 320px;
-        margin: 0 0px;
-    }
 
-    .articleImg {
-        width: 100%;
-        /* min-width: 320px; */
-        height: 100%;
-        /* min-height: 320px; */
-        margin: 0 0px;
-    }
-    .imgBar {
-        display: none;
-    }
-    .imgBarMob {
-        display: block;
-        width: 68px;
-        height: 8px;
-        position: absolute;
-        bottom: 20px;
-        left: calc(50% - (68px / 2));
-        display: -webkit-box;
-        display: -ms-flexbox;
-        display: flex;
-        -webkit-box-pack: justify;
-        -ms-flex-pack: justify;
-        justify-content: space-between;
-        cursor: pointer;
-    }
     .articleRight {
         max-width: 100%;
         width: 100%;
         padding: 0 20px;
     }
-    .articleTitle {
-        margin-bottom: 10px;
-    }
-    .articleInfo {
-        margin-bottom: 20px;
-    }
-    .articleDate,
-    .articleCity {
-        font-size: 14px;
-        line-height: 17px;
-        color: #5f5f5f;
-        margin-bottom: 4px;
-    }
-    .articleLink {
-        font-size: 14px;
-        line-height: 19px;
-        color: #009ee4;
-    }
-    .articlePrice {
-        font-size: 18px;
-        line-height: 25px;
-        font-weight: 700;
-        margin-bottom: 20px;
-    }
-    .articleBtn {
-        width: 100%;
-        height: 57px;
-        font-size: 14px;
-        line-height: 20px;
-        color: #ffffff;
-    }
-
-    .articleAuthor {
-        margin-top: 30px;
-        margin-bottom: 30px;
-    }
-    .imgBarMobCircle {
-        width: 8px;
-        height: 8px;
-        border-radius: 50%;
-        border: 1px solid #009ee4;
-        transition: all 0.3s;
-    }
-    .imgBarMobCircle:hover {
-        border: 1px solid #0080c1;
-        background-color: #0080c1;
-    }
-    .imgBarMobCircle:active {
-        border: 1px solid #0080c1;
-        background-color: #0080c1;
-    }
-    .authorName {
-        font-size: 18px;
-        line-height: 23px;
-        font-weight: 600;
-    }
-    .authorAbout {
-        font-size: 14px;
-        line-height: 28px;
-    }
-}
-
-@media screen and (max-width: 580px) {
-    .articleFillImg {
-        max-width: 100%;
-        width: 100%;
-        overflow: hidden;
-        display: -webkit-box;
-        display: -ms-flexbox;
-        display: flex;
-        -webkit-box-orient: vertical;
-        -webkit-box-direction: normal;
-        -ms-flex-direction: column;
-        flex-direction: column;
-        -webkit-box-align: center;
-        -ms-flex-align: center;
-        align-items: center;
-        -webkit-box-pack: center;
-        -ms-flex-pack: center;
-        justify-content: center;
-        position: relative;
-    }
-    /* .articleFillImg::before {
-        content: '';
-        display: block;
-        width: 23px;
-        height: 23px;
-        background-color: transparent;
-        border-top: 2px solid #000000;
-        border-left: 2px solid #000000;
-        -webkit-transform: rotate(-45deg);
-        transform: rotate(-45deg);
-        position: absolute;
-        top: 24px;
-        left: 32px;
-        cursor: pointer;
-    } */
 }
